Demo both versions of ct-checkbox
